# docs-SuMon/OpenApi-SuMon/openapi-server.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SuSuMonitor Server Management API",
    "description": "Authoritative contract for secure server management. JSON property names use snake_case. Credential plaintext and ciphertext are request-only implementation concerns and are never returned by any response.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:18080",
      "description": "Local development server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/servers": {
      "get": {
        "tags": ["servers"],
        "summary": "List servers",
        "description": "Available to any authenticated user. Returns non-deleted servers without credential plaintext or ciphertext.",
        "operationId": "listServers",
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/PageSize"},
          {"$ref": "#/components/parameters/Keyword"},
          {"$ref": "#/components/parameters/SortBy"},
          {"$ref": "#/components/parameters/SortOrder"},
          {"$ref": "#/components/parameters/XCorrelationId"}
        ],
        "responses": {
          "200": {
            "description": "Server page returned",
            "headers": {
              "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
              "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ServerPageResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequestParameter"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      },
      "post": {
        "tags": ["servers"],
        "summary": "Create server",
        "description": "Admin only. The primary credential matching ssh_auth_type is required. password and private_key are mutually exclusive, blank credentials are invalid, and responses never contain credential plaintext or ciphertext.",
        "operationId": "createServer",
        "parameters": [
          {"$ref": "#/components/parameters/XCorrelationId"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateServerRequest"},
              "examples": {
                "passwordAuthentication": {
                  "value": {
                    "name": "production-web-01",
                    "host": "web-01.example.com",
                    "description": "Production web server",
                    "ssh_host": "web-01.example.com",
                    "ssh_port": 22,
                    "ssh_user": "monitor",
                    "ssh_auth_type": "password",
                    "ssh_password": "REPLACE_WITH_PASSWORD"
                  }
                },
                "privateKeyAuthentication": {
                  "value": {
                    "name": "production-db-01",
                    "host": "db-01.example.com",
                    "description": "Production database server",
                    "ssh_host": "db-01.example.com",
                    "ssh_port": 22,
                    "ssh_user": "monitor",
                    "ssh_auth_type": "private_key",
                    "ssh_private_key": "REPLACE_WITH_PRIVATE_KEY"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server created",
            "headers": {
              "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
              "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ServerResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequestParameter"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/servers/{id}": {
      "get": {
        "tags": ["servers"],
        "summary": "Get server",
        "description": "Available to any authenticated user. Returns a non-deleted server without credential plaintext or ciphertext.",
        "operationId": "getServer",
        "parameters": [
          {"$ref": "#/components/parameters/ServerId"},
          {"$ref": "#/components/parameters/XCorrelationId"}
        ],
        "responses": {
          "200": {
            "description": "Server returned",
            "headers": {
              "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
              "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ServerResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequestParameter"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      },
      "put": {
        "tags": ["servers"],
        "summary": "Update server",
        "description": "Admin only. All basic fields are required. Omitting password and private_key retains the stored primary credential when ssh_auth_type is unchanged. Blank credentials and submitting both credential types are invalid. Changing ssh_auth_type requires the new matching primary credential. Responses never contain credential plaintext or ciphertext.",
        "operationId": "updateServer",
        "parameters": [
          {"$ref": "#/components/parameters/ServerId"},
          {"$ref": "#/components/parameters/XCorrelationId"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/UpdateServerRequest"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server updated",
            "headers": {
              "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
              "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ServerResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequestParameter"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      },
      "delete": {
        "tags": ["servers"],
        "summary": "Delete server",
        "description": "Admin only. Soft-deletes the server; no credential plaintext or ciphertext is returned.",
        "operationId": "deleteServer",
        "parameters": [
          {"$ref": "#/components/parameters/ServerId"},
          {"$ref": "#/components/parameters/XCorrelationId"}
        ],
        "responses": {
          "200": {
            "description": "Server deleted",
            "headers": {
              "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
              "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/EmptyResponse"},
                "example": {"code": 0, "message": "success", "data": null}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequestParameter"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/servers/{id}/status": {
      "get": {
        "tags": ["servers"],
        "summary": "Get server status snapshot",
        "description": "Available to any authenticated user. Returns the latest status snapshot stored in the database. It does not trigger and must not be interpreted as a real-time probe.",
        "operationId": "getServerStatus",
        "parameters": [
          {"$ref": "#/components/parameters/ServerId"},
          {"$ref": "#/components/parameters/XCorrelationId"}
        ],
        "responses": {
          "200": {
            "description": "Stored server status snapshot returned",
            "headers": {
              "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
              "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ServerStatusResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequestParameter"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalServerError"}
        }
      }
    },
    "/api/servers/{id}/ssh/host-key": {
      "put": {
        "tags": ["servers"],
        "summary": "Confirm or rotate SSH host key",
        "description": "Admin only. Performs target validation and an SSH handshake to compare the observed host key with expected_fingerprint, but never sends password, private key, passphrase, or any other login credential. A first confirmation stores the key, replace=true explicitly rotates a different stored key, and confirming the currently stored key is an idempotent review.",
        "operationId": "confirmServerSshHostKey",
        "x-required-role": "admin",
        "parameters": [
          {"$ref": "#/components/parameters/ServerId"},
          {"$ref": "#/components/parameters/XCorrelationId"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ConfirmSshHostKeyRequest"},
              "example": {
                "expected_fingerprint": "SHA256:REPLACEWITHBASE64FINGERPRINT",
                "replace": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Host key confirmed, rotated, or reviewed without change",
            "headers": {
              "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
              "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/SshHostKeyResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequestParameter"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/SshForbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/SshHostKeyConflict"},
          "429": {"$ref": "#/components/responses/SshConnectionLimitExceeded"},
          "500": {"$ref": "#/components/responses/DatabaseError"},
          "502": {"$ref": "#/components/responses/SshConnectionFailed"},
          "504": {"$ref": "#/components/responses/SshTimeout"}
        }
      }
    },
    "/api/servers/{id}/ssh/test": {
      "post": {
        "tags": ["servers"],
        "summary": "Test SSH authentication",
        "description": "Admin only. Accepts no request body. Requires an already registered host key, validates the handshake against that exact key, and then tests the server's stored password or private_key credential. It never registers, replaces, or updates a host key and closes the connection without executing a command or creating a PTY.",
        "operationId": "testServerSshConnection",
        "x-required-role": "admin",
        "parameters": [
          {"$ref": "#/components/parameters/ServerId"},
          {"$ref": "#/components/parameters/XCorrelationId"}
        ],
        "responses": {
          "200": {
            "description": "SSH host key and stored credential verified",
            "headers": {
              "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
              "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/SshConnectionTestResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequestParameter"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/SshForbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/SshHostKeyConflict"},
          "429": {"$ref": "#/components/responses/SshConnectionLimitExceeded"},
          "500": {"$ref": "#/components/responses/DatabaseError"},
          "502": {"$ref": "#/components/responses/SshConnectionFailed"},
          "504": {"$ref": "#/components/responses/SshTimeout"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT supplied only through the Authorization: Bearer <token> request header."
      }
    },
    "parameters": {
      "ServerId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Server ID.",
        "schema": {"type": "integer", "format": "int64", "minimum": 1}
      },
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "One-based page number.",
        "schema": {"type": "integer", "format": "int32", "minimum": 1, "default": 1}
      },
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "description": "Number of items per page.",
        "schema": {"type": "integer", "format": "int32", "minimum": 1, "maximum": 100, "default": 20}
      },
      "Keyword": {
        "name": "keyword",
        "in": "query",
        "required": false,
        "description": "Optional keyword matched against server name, host or description.",
        "schema": {"type": "string", "minLength": 1, "maxLength": 100}
      },
      "SortBy": {
        "name": "sort_by",
        "in": "query",
        "required": false,
        "description": "Sort field whitelist.",
        "schema": {
          "type": "string",
          "enum": ["id", "name", "host", "status", "created_at", "updated_at"],
          "default": "id"
        }
      },
      "SortOrder": {
        "name": "sort_order",
        "in": "query",
        "required": false,
        "description": "Sort direction.",
        "schema": {"type": "string", "enum": ["asc", "desc"], "default": "desc"}
      },
      "XCorrelationId": {
        "name": "X-Correlation-ID",
        "in": "header",
        "required": false,
        "description": "Optional client correlation ID. Invalid values are ignored and never logged.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 64,
          "pattern": "^[A-Za-z0-9_-]+$"
        }
      }
    },
    "headers": {
      "XRequestId": {
        "description": "UUID request ID generated by the server. Client-provided X-Request-ID values are ignored.",
        "schema": {"type": "string", "format": "uuid", "readOnly": true}
      },
      "XCorrelationId": {
        "description": "Optional validated client correlation ID, returned only when a valid request header was provided.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 64,
          "pattern": "^[A-Za-z0-9_-]+$"
        }
      }
    },
    "schemas": {
...
      "CreateServerRequest": {
        "description": "Create request. Exactly one matching primary credential is required; submitting both types is invalid.",
        "allOf": [
          {"$ref": "#/components/schemas/ServerBasicFields"},
          {"$ref": "#/components/schemas/CredentialFields"},
          {
            "anyOf": [
              {
                "required": ["ssh_auth_type", "ssh_password"],
                "properties": {"ssh_auth_type": {"type": "string", "enum": ["password"]}}
              },
              {
                "required": ["ssh_auth_type", "ssh_private_key"],
                "properties": {"ssh_auth_type": {"type": "string", "enum": ["private_key"]}}
              }
            ]
          }
        ]
      },
      "UpdateServerRequest": {
        "description": "All basic fields are required. Credentials may be omitted only to retain the existing credential when ssh_auth_type is unchanged. If ssh_auth_type changes, the service must require the matching new primary credential. Submitting both credential types or an empty credential is invalid.",
        "allOf": [
          {"$ref": "#/components/schemas/ServerBasicFields"},
          {"$ref": "#/components/schemas/CredentialFields"},
          {
            "required": ["description"],
            "anyOf": [
              {
                "required": ["ssh_auth_type"],
                "properties": {"ssh_auth_type": {"type": "string", "enum": ["password"]}}
              },
              {
                "required": ["ssh_auth_type"],
                "properties": {"ssh_auth_type": {"type": "string", "enum": ["private_key"]}}
              }
            ]
          }
        ]
      },
...
      "ServerResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/ApiResponse"},
          {
            "type": "object",
            "properties": {"data": {"$ref": "#/components/schemas/Server"}}
          }
        ]
      },
      "ServerPageResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/ApiResponse"},
          {
            "type": "object",
            "properties": {"data": {"$ref": "#/components/schemas/ServerPageResult"}}
          }
        ]
      },
      "ServerStatusResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/ApiResponse"},
          {
            "type": "object",
            "properties": {"data": {"$ref": "#/components/schemas/ServerStatus"}}
          }
        ]
      },
      "ConfirmSshHostKeyRequest": {
        "type": "object",
        "required": ["expected_fingerprint"],
        "properties": {
          "expected_fingerprint": {
            "type": "string",
            "minLength": 50,
            "maxLength": 50,
            "pattern": "^SHA256:[A-Za-z0-9+/]{43}$",
            "description": "Expected OpenSSH SHA-256 host key fingerprint verified by the administrator through a trusted out-of-band channel."
          },
          "replace": {
            "type": "boolean",
            "default": false,
            "description": "Must be true to replace a different registered fingerprint after an administrator verifies a legitimate host key rotation."
          }
        }
      },
...
      "SshHostKeyResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/ApiResponse"},
          {
            "type": "object",
            "properties": {"data": {"$ref": "#/components/schemas/SshHostKeyResult"}}
          }
        ]
      },
...
      "EmptyResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/ApiResponse"},
          {
            "type": "object",
            "properties": {"data": {"nullable": true, "example": null}}
          }
        ]
      },
...
    },
    "responses": {
      "InvalidRequestParameter": {
        "description": "Invalid request parameter or credential combination",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "example": {"code": 40002, "message": "invalid request parameter", "data": null}
          }
        }
      },
      "Unauthorized": {
        "description": "Bearer JWT is missing, invalid, expired, or no longer authorized",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "example": {"code": 40100, "message": "unauthorized", "data": null}
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated user is not an admin",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "example": {"code": 40300, "message": "forbidden", "data": null}
          }
        }
      },
      "SshForbidden": {
        "description": "The caller is not an admin or the SSH target violates the outbound access policy",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "examples": {
              "adminRequired": {"value": {"code": 40300, "message": "forbidden", "data": null}},
              "targetForbidden": {"value": {"code": 40301, "message": "ssh target forbidden", "data": null}}
            }
          }
        }
      },
      "NotFound": {
        "description": "Server does not exist or has been deleted",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "example": {"code": 40400, "message": "resource not found", "data": null}
          }
        }
      },
      "Conflict": {
        "description": "Server name, host, concurrent state, or delete constraint conflicts with the request",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "example": {"code": 40900, "message": "resource conflict", "data": null}
          }
        }
      },
      "SshHostKeyConflict": {
        "description": "The host key is not registered, does not match, or changed concurrently",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "examples": {
              "concurrentUpdate": {"value": {"code": 40900, "message": "resource conflict", "data": null}},
              "notConfirmed": {"value": {"code": 40901, "message": "ssh host key not confirmed", "data": null}},
              "fingerprintMismatch": {"value": {"code": 40902, "message": "ssh host key mismatch", "data": null}}
            }
          }
        }
      },
      "SshConnectionLimitExceeded": {
        "description": "SSH connection test concurrency limit reached",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "example": {"code": 42900, "message": "ssh connection limit reached", "data": null}
          }
        }
      },
      "DatabaseError": {
        "description": "Database operation failed",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "example": {"code": 50001, "message": "database error", "data": null}
          }
        }
      },
      "SshConnectionFailed": {
        "description": "The SSH handshake or stored credential authentication could not be completed",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "examples": {
              "connectionFailed": {"value": {"code": 50002, "message": "ssh connection failed", "data": null}},
              "authenticationFailed": {"value": {"code": 50003, "message": "ssh authentication failed", "data": null}}
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server or database error",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/XRequestId"},
          "X-Correlation-ID": {"$ref": "#/components/headers/XCorrelationId"}
        },
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"},
            "examples": {
              "internalError": {
                "value": {"code": 50000, "message": "internal server error", "data": null}
              },
              "databaseError": {
                "value": {"code": 50001, "message": "database error", "data": null}
              }
            }
          }
        }
      }
    }
  }
}